[2018-02-06] Challenge #350 [Easy] Bookshelf problem
This turned out to be a little harder than I thought.
Mostly due to my misunderstanding of the SortedSet data structure in .NET.
I attempted to use the SortedSet for the book collections, and I created a custom comparer for the ordering. This should have worked because the struct had its own hash function, but the sorted set seemed to consider data with the same width as duplicates.
I assume this is because of the comparer I passed in, which left me without a way of telling the data structure how to sort independent of how to identify uniqueness.
The resulting solution is inefficient in its use of data structures but it's getting late and i'm tired.